+++ GoogleAdManager/src/main/java/com/applovin/mediation/adapters/GoogleAdManagerMediationAdapter.java
@@ -237,7 +237,9 @@ public class GoogleAdManagerMediationAdapter
         {
             log( "Interstitial ad failed to show: " + placementId );
 
-            MaxAdapterError error = new MaxAdapterError( MaxAdapterError.AD_DISPLAY_FAILED, 0, "Interstitial ad not ready" );
+            MaxAdapterError error = new MaxAdapterError( MaxAdapterError.AD_DISPLAY_FAILED,
+                                                         MaxAdapterError.AD_NOT_READY.getCode(),
+                                                         MaxAdapterError.AD_NOT_READY.getMessage() );
             listener.onInterstitialAdDisplayFailed( error );
         }
     }
@@ -307,7 +309,9 @@ public class GoogleAdManagerMediationAdapter
         {
             log( "App open ad failed to show: " + placementId );
 
-            MaxAdapterError error = new MaxAdapterError( MaxAdapterError.AD_DISPLAY_FAILED, 0, "App open ad not ready" );
+            MaxAdapterError error = new MaxAdapterError( MaxAdapterError.AD_DISPLAY_FAILED,
+                                                         MaxAdapterError.AD_NOT_READY.getCode(),
+                                                         MaxAdapterError.AD_NOT_READY.getMessage() );
             listener.onAppOpenAdDisplayFailed( error );
         }
     }
@@ -383,7 +387,9 @@ public class GoogleAdManagerMediationAdapter
         {
             log( "Rewarded ad failed to show: " + placementId );
 
-            MaxAdapterError error = new MaxAdapterError( MaxAdapterError.AD_DISPLAY_FAILED, 0, "Rewarded ad not ready" );
+            MaxAdapterError error = new MaxAdapterError( MaxAdapterError.AD_DISPLAY_FAILED,
+                                                         MaxAdapterError.AD_NOT_READY.getCode(),
+                                                         MaxAdapterError.AD_NOT_READY.getMessage() );
             listener.onRewardedAdDisplayFailed( error );
         }
     }
@@ -760,7 +766,9 @@ public class GoogleAdManagerMediationAdapter
         @Override
         public void onAdFailedToShowFullScreenContent(@NonNull final AdError adError)
         {
-            MaxAdapterError adapterError = new MaxAdapterError( MaxAdapterError.AD_DISPLAY_FAILED, adError.getCode(), adError.getMessage() );
+            MaxAdapterError adapterError = new MaxAdapterError( MaxAdapterError.AD_DISPLAY_FAILED,
+                                                                adError.getCode(),
+                                                                adError.getMessage() );
             log( "Interstitial ad (" + placementId + ") failed to show with error: " + adapterError );
             listener.onInterstitialAdDisplayFailed( adapterError );
         }
@@ -808,7 +816,9 @@ public class GoogleAdManagerMediationAdapter
         @Override
         public void onAdFailedToShowFullScreenContent(@NonNull final AdError adError)
         {
-            MaxAdapterError adapterError = new MaxAdapterError( MaxAdapterError.AD_DISPLAY_FAILED, adError.getCode(), adError.getMessage() );
+            MaxAdapterError adapterError = new MaxAdapterError( MaxAdapterError.AD_DISPLAY_FAILED,
+                                                                adError.getCode(),
+                                                                adError.getMessage() );
             log( "App open ad (" + placementId + ") failed to show with error: " + adapterError );
             listener.onAppOpenAdDisplayFailed( adapterError );
         }
@@ -858,7 +868,9 @@ public class GoogleAdManagerMediationAdapter
         @Override
         public void onAdFailedToShowFullScreenContent(@NonNull final AdError adError)
         {
-            MaxAdapterError adapterError = new MaxAdapterError( MaxAdapterError.AD_DISPLAY_FAILED, adError.getCode(), adError.getMessage() );
+            MaxAdapterError adapterError = new MaxAdapterError( MaxAdapterError.AD_DISPLAY_FAILED,
+                                                                adError.getCode(),
+                                                                adError.getMessage() );
             log( "Rewarded ad (" + placementId + ") failed to show with error: " + adapterError );
             listener.onRewardedAdDisplayFailed( adapterError );
         }